Rust nigthly & dependency fixes (#15)
* Add rust-toolchain.toml file

This file tells cargo (and other tools, like vscode rust-analyzer
plugin) to use the nightly toolchain by default.

This is needed to correctly open the project in vscode with
rust-analyzer.

We should probably fix a nigthly version, but for now leave it as a
generic `nightly`.

* Use feature(generic_associated_types) as in stable

The feature(generic_associated_types) has been recently stabilized and
does not need to be listed in the `#![feature(...)]` block.

Also, move the `where` clauses on GAT impls to after the type
assignment, taking advantage of a syntax change described here:

  https://github.com/rust-lang/rust/issues/89122

* Update bitvec to 1.0.1

bitvec-0.22.3 depends on yanked funty-1.2.0 and no longer builds. Update
to bitvec-1.0.1 as this is the most recent version.

Change a few API calls as they have been renamed. Move the framebuffer
BitArray size to a constant so that it is not repeated.

* Fix a typo in settings.toml

* Invoke `cargo fmt` on the whole workspace

* Fix apex-ctl clap usage and update it

Update clap to 4.0.26. Enable the `derive` feature to fix compilation.

Change the derive attributes to work with newer version of clap.

Drop apex_hardware::FareBuffer use.

* Fix `debug` feature

Make `src/render/debug.rs` buildable.

* Fix `coindesk` section in `settings.toml`

It was called `crypto`, but the provider is called `coindesk` instead.

* fixup! Update bitvec to 1.0.1